+++ examples/simple.rs
@@ -4,11 +4,11 @@ fn main() -> Result<(), Error> {
     let mut tile = Tile::new(4096);
     let layer = tile.create_layer("First Layer");
     let b = GeomEncoder::new(GeomType::Linestring, Transform::new())
-                        .add_point(0.0, 0.0)
-                        .add_point(1024.0, 0.0)
-                        .add_point(1024.0, 2048.0)
-                        .add_point(2048.0, 2048.0)
-                        .add_point(2048.0, 4096.0)
+                        .point(0.0, 0.0)
+                        .point(1024.0, 0.0)
+                        .point(1024.0, 2048.0)
+                        .point(2048.0, 2048.0)
+                        .point(2048.0, 4096.0)
                         .encode()?;
     let mut feature = layer.into_feature(b);
     feature.set_id(1)?;