+++ src/tile.rs
@@ -73,8 +73,8 @@ pub struct Layer {
 ///       let mut tile = Tile::new(4096);
 ///       let layer = tile.create_layer("First Layer");
 ///       let geom_data = GeomEncoder::new(GeomType::Point, Transform::new())
-///                                   .add_point(1.0, 2.0)
-///                                   .add_point(7.0, 6.0)
+///                                   .point(1.0, 2.0)
+///                                   .point(7.0, 6.0)
 ///                                   .encode()?;
 ///       let feature = layer.into_feature(geom_data);
 ///       // ...